refactor move schema class apis/users to schema/users
